Fix various compiler warnings (W4) and refactor code for C++20 compliance:
Fix C4505: Changed header-only helper functions from 'static' to 'inline'.
* Fix C4244: Added 'static_cast' and 'decltype' to resolve data truncation and loss warnings.
* Fix C4100: Applied '[[maybe_unused]]' to version-dependent function parameters.
* Fix C4456/C4458: Resolved multiple variable shadowing warnings (e.g., 'guid', 'pck', structured bindings) by renaming or removing redundant local declarations.
* Fix MoP Packet Bug: Fixed a critical logic bug caused by shadowing 'guid' in 'internalDeserialise', which previously led to data loss.
* Fix C4700: Initialized 'spellId' to 0 in cancel-cast opcode to eliminate uninitialized variable usage and undefined behavior.
* Fix C5054: Cast distinct enums to 'uint32_t' before addition to satisfy strict C++20 enum arithmetic rules.
* Fix C4189: Cleaned up unreferenced initialized variables and removed redundant duplicate function calls (e.g., 'getVehicleBase').
* Code Safety: Replaced unsafe static downcasts on 'pCorpse' with secure 'dynamic_cast' and 'nullptr' validation.
* Memory Alignment: Resolved unary '&' operator warnings on packed network structures by using aligned temporary stack variables before byteswapping.
* Bitstream Corruption: Removed a duplicate, bracketless 'for' loop that corrupted the aura update network stream.
* Modernization: Replaced legacy C-style casts with typsicher 'static_cast' and 'reinterpret_cast', minimized variable scopes, and updated boolean flags.

+++ src/scripts/Battlegrounds/EyeOfTheStorm/EyeOfTheStorm.cpp
@@ -329,7 +329,7 @@ void EyeOfTheStorm::HookOnAreaTrigger(Player* plr, uint32_t id)
             break;
     }
 
-    if (plr->isDead())        // on ne buff pas les joueurs morts ;)
+    if (plr->isDead()) // on ne buff pas les joueurs morts ;)
         return;
 
     if (bonusid > -1)
@@ -359,7 +359,7 @@ void EyeOfTheStorm::HookOnAreaTrigger(Player* plr, uint32_t id)
         sendChatMessage(CHAT_MSG_BG_EVENT_NEUTRAL, plr->getGuid(), "%s will be removed from the game for cheating.", plr->getName().c_str());
         // Remove player from battleground.
         this->removePlayer(plr, false);
-        // Kick    player from server.
+        // Kick player from server.
         plr->kickFromServer(TimeVarsMs::Second * 6);
         return;
     }
@@ -377,7 +377,7 @@ void EyeOfTheStorm::HookOnAreaTrigger(Player* plr, uint32_t id)
         val = EOTS_BANNER_HORDE;
 
     if (!m_CPBanner[tid] || m_CPBanner[tid]->getEntry() != static_cast<uint32_t>(val))
-        return;            // not captured by our team
+        return; // not captured by our team
 
     for (uint8_t i = 0; i < EOTS_TOWER_COUNT; ++i)
     {
@@ -387,17 +387,17 @@ void EyeOfTheStorm::HookOnAreaTrigger(Player* plr, uint32_t id)
 
     /*
     Points from flag captures
-    * 1 towers controlled = 75 points
+    * 1 tower controlled = 75 points
     * 2 towers controlled = 85 points
     * 3 towers controlled = 100 points
     * 4 towers controlled = 500 points
     */
 
     // 25 is guessed
-    const static uint32_t points[5] = { 25, 75, 85, 100, 500 };
-    const char* msgs[2] = { "The Alliance have captured the flag.", "The Horde have captured the flag." };
+    const static uint32_t points[5] = {25, 75, 85, 100, 500};
+    const char* msgs[2] = {"The Alliance have captured the flag.", "The Horde have captured the flag."};
 
-    sendChatMessage(CHAT_MSG_BG_EVENT_ALLIANCE + team, 0, msgs[team]);
+    sendChatMessage(static_cast<uint8_t>(CHAT_MSG_BG_EVENT_ALLIANCE) + static_cast<uint8_t>(team), 0, msgs[team]);
     GivePoints(team, points[towers]);
 
     DropFlag2(plr, id);
@@ -434,7 +434,7 @@ void EyeOfTheStorm::HookFlagDrop(Player* plr, GameObject* /*obj*/)
 
     setWorldState(EOTS_NETHERWING_FLAG_READY, 0);
     playSoundToAll(plr->isTeamHorde() ? BattlegroundDef::HORDE_CAPTURE : BattlegroundDef::ALLIANCE_CAPTURE);
-    sendChatMessage(CHAT_MSG_BG_EVENT_ALLIANCE + plr->getTeam(), plr->getGuid(), "%s has taken the flag!", plr->getName().c_str());
+    sendChatMessage(static_cast<uint8_t>(CHAT_MSG_BG_EVENT_ALLIANCE) + static_cast<uint8_t>(plr->getTeam()), plr->getGuid(), "%s has taken the flag!", plr->getName().c_str());
     m_flagHolder = plr->getGuidLow();
 
     event_RemoveEvents(EVENT_EOTS_RESET_FLAG);
@@ -454,7 +454,7 @@ bool EyeOfTheStorm::HookSlowLockOpen(GameObject* /*pGo*/, Player* pPlayer, Spell
 
     setWorldState(EOTS_NETHERWING_FLAG_READY, 0);
     playSoundToAll(pPlayer->isTeamHorde() ? BattlegroundDef::HORDE_CAPTURE : BattlegroundDef::ALLIANCE_CAPTURE);
-    sendChatMessage(CHAT_MSG_BG_EVENT_ALLIANCE + pPlayer->getTeam(), pPlayer->getGuid(), "%s has taken the flag!", pPlayer->getName().c_str());
+    sendChatMessage(static_cast<uint8_t>(CHAT_MSG_BG_EVENT_ALLIANCE) + static_cast<uint8_t>(pPlayer->getTeam()), pPlayer->getGuid(), "%s has taken the flag!", pPlayer->getName().c_str());
     m_flagHolder = pPlayer->getGuidLow();
     return true;
 }
@@ -537,7 +537,7 @@ void EyeOfTheStorm::HookOnFlagDrop(Player* plr)
     m_dropFlag->PushToWorld(m_mapMgr);
     m_flagHolder = 0;
     playSoundToAll(BattlegroundDef::FLAG_RETURNED);
-    sendChatMessage(CHAT_MSG_BG_EVENT_ALLIANCE + plr->getTeam(), plr->getGuid(), "%s has dropped the flag!", plr->getName().c_str());
+    sendChatMessage(static_cast<uint8_t>(CHAT_MSG_BG_EVENT_ALLIANCE) + static_cast<uint8_t>(plr->getTeam()), plr->getGuid(), "%s has dropped the flag!", plr->getName().c_str());
 
     sEventMgr.AddEvent(this, &EyeOfTheStorm::EventResetFlag, EVENT_EOTS_RESET_FLAG, 10000, 1, EVENT_FLAG_DO_NOT_EXECUTE_IN_WORLD_CONTEXT);
 }
